add lucene analyzer constants
