+++ textdb/textdb-dataflow/src/main/java/edu/uci/ics/textdb/plangen/operatorbuilder/DictionaryMatcherBuilder.java
@@ -4,8 +4,8 @@ import java.util.List;
 import java.util.Map;
 
 import edu.uci.ics.textdb.api.common.Attribute;
-import edu.uci.ics.textdb.common.constants.DataConstants;
 import edu.uci.ics.textdb.common.constants.DataConstants.KeywordMatchingType;
+import edu.uci.ics.textdb.common.constants.LuceneAnalyzerConstants;
 import edu.uci.ics.textdb.common.exception.DataFlowException;
 import edu.uci.ics.textdb.common.exception.PlanGenException;
 import edu.uci.ics.textdb.dataflow.common.Dictionary;
@@ -53,7 +53,7 @@ public class DictionaryMatcherBuilder {
 
         // build DictionaryMatcher
         DictionaryPredicate predicate = new DictionaryPredicate(dictionary, attributeList,
-                DataConstants.getStandardAnalyzer(), matchingType);
+                LuceneAnalyzerConstants.getStandardAnalyzer(), matchingType);
         DictionaryMatcher operator = new DictionaryMatcher(predicate);
 
         // set limit and offset
