+++ textdb/textdb-dataflow/src/main/java/edu/uci/ics/textdb/plangen/operatorbuilder/KeywordSourceBuilder.java
@@ -4,8 +4,8 @@ import java.util.List;
 import java.util.Map;
 
 import edu.uci.ics.textdb.api.common.Attribute;
-import edu.uci.ics.textdb.common.constants.DataConstants;
 import edu.uci.ics.textdb.common.constants.DataConstants.KeywordMatchingType;
+import edu.uci.ics.textdb.common.constants.LuceneAnalyzerConstants;
 import edu.uci.ics.textdb.common.exception.DataFlowException;
 import edu.uci.ics.textdb.common.exception.PlanGenException;
 import edu.uci.ics.textdb.dataflow.common.KeywordPredicate;
@@ -52,7 +52,7 @@ public class KeywordSourceBuilder {
         KeywordPredicate keywordPredicate;
         try {
             keywordPredicate = new KeywordPredicate(keyword, attributeList,
-                    DataConstants.getStandardAnalyzer(), matchingType);     
+                    LuceneAnalyzerConstants.getStandardAnalyzer(), matchingType);     
         } catch (DataFlowException e) {
             throw new PlanGenException(e.getMessage(), e);
         }
