+++ textdb/textdb-dataflow/src/main/java/edu/uci/ics/textdb/plangen/operatorbuilder/RegexMatcherBuilder.java
@@ -5,7 +5,7 @@ import java.util.List;
 import java.util.Map;
 
 import edu.uci.ics.textdb.api.common.Attribute;
-import edu.uci.ics.textdb.common.constants.DataConstants;
+import edu.uci.ics.textdb.common.constants.LuceneAnalyzerConstants;
 import edu.uci.ics.textdb.common.exception.DataFlowException;
 import edu.uci.ics.textdb.common.exception.PlanGenException;
 import edu.uci.ics.textdb.dataflow.common.RegexPredicate;
@@ -43,7 +43,7 @@ public class RegexMatcherBuilder {
         RegexPredicate regexPredicate;
         try {
             regexPredicate = new RegexPredicate(regex, attributeList,
-                    DataConstants.getNGramAnalyzer(3));
+                    LuceneAnalyzerConstants.getNGramAnalyzer(3));
         } catch (DataFlowException e) {
             throw new PlanGenException(e.getMessage(), e);
         }
