+++ textdb/textdb-perftest/src/main/java/edu/uci/ics/textdb/perftest/utils/PerfTestUtils.java
@@ -13,7 +13,7 @@ import java.util.Scanner;
 import org.apache.lucene.analysis.Analyzer;
 import org.apache.lucene.analysis.standard.StandardAnalyzer;
 
-import edu.uci.ics.textdb.common.constants.DataConstants;
+import edu.uci.ics.textdb.common.constants.LuceneAnalyzerConstants;
 import edu.uci.ics.textdb.engine.Engine;
 import edu.uci.ics.textdb.perftest.medline.MedlineIndexWriter;
 import edu.uci.ics.textdb.storage.DataStore;
@@ -182,7 +182,7 @@ public class PerfTestUtils {
             if (file.isDirectory()) {
                 continue;
             }
-            writeIndex(file.getName(), DataConstants.getNGramAnalyzer(3), "trigram");
+            writeIndex(file.getName(), LuceneAnalyzerConstants.getNGramAnalyzer(3), "trigram");
         }
 
     }
